Refactor command handlers with shared base
